Decomp `ActorRefill` 94%

+++ include/Item/ItemManager.hpp
@@ -134,7 +134,7 @@ public:
 
     // Ammo
     u16 GetAmmo(ItemFlag equipId) const;
-    void GiveAmmo(ItemFlag equipId, u16 amount);
+    void GiveAmmo(ItemFlag equipId, u32 amount);
     u16 GetMaxAmmo(ItemFlag equipId) const;
     void UpgradeQuiver();
     void UpgradeBombBag();
